requirements.txt
uvicorn~=0.34.0
fastapi~=0.115.11
python-dotenv~=1.0.1
pydantic~=2.10.6
kafka-python~=2.0.6
